Service And Repository

+++ workspace/ims-webapp/src/main/java/th/co/baiwa/buckwaframework/accesscontrol/persistence/repository/AdmRoleOperationRepository.java
@@ -1,13 +1,10 @@
 
 package th.co.baiwa.buckwaframework.accesscontrol.persistence.repository;
 
-import java.math.BigDecimal;
 import th.co.baiwa.buckwaframework.accesscontrol.persistence.entity.AdmRoleOperation;
 import th.co.baiwa.buckwaframework.common.persistence.repository.CommonJpaCrudRepository;
 
-public interface AdmRoleOperationRepository
-    extends CommonJpaCrudRepository<AdmRoleOperation, BigDecimal>
-{
-
+public interface AdmRoleOperationRepository extends CommonJpaCrudRepository<AdmRoleOperation, Long> {
 
 }
+
